pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-agent-orchestrator"
version = "0.1.0"
description = "AI-driven dev orchestrator for FastAPI codebases: docs, tests, and bug fixes. Supports Anthropic and OpenAI."
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.50.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.40",
    "jinja2>=3.1.0",
    "watchfiles>=0.21.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[project.scripts]
orchestrator = "orchestrator.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
